pkg: Move existing GCS implementation to the `object` package
